myWeatherForecast/requirements.txt
fastapi
uvicorn
requests
apscheduler
python-dotenv
pydantic
httpx
pytest
